Add Search Functionality And Screen
The new screen is needed to allow users to find a issue by release
number.

The change in the PublicationBloc was needed to allow more listeners on
the result stream and to cache the fetched value because the same data
is used in the search process and can be requested at any time.